requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
distro==1.8.0
Django==5.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.8
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
whitenoise==6.6.0
